+++ src/config/chain/initia.ts
@@ -8,14 +8,14 @@ const INITIA_DECODER =
   "https://initia-api-jiod42ec2q-as.a.run.app/decode_module";
 
 export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
-  "tomcat-1-lite": {
+  "minimove-1-lite": {
     tier: "lite",
     chain: "initia",
-    registryChainName: "blackwingtestnet",
-    prettyName: "Blackwing Testnet Lite",
-    lcd: "https://maze-rest-18bdff44-3aa4-425e-9bc0-06a2afa40af8.ase1-prod.newmetric.xyz",
-    rpc: "https://maze-rpc-18bdff44-3aa4-425e-9bc0-06a2afa40af8.ase1-prod.newmetric.xyz",
-    indexer: "https://tomcat-1-graphql.alleslabs.dev/v1/graphql",
+    registryChainName: "minimovetestnet",
+    prettyName: "Minimove",
+    lcd: "https://lcd.minimove-1.initia.xyz",
+    rpc: "https://rpc.minimove-1.initia.xyz",
+    indexer: "https://minimove-1-graphql.alleslabs.dev/v1/graphql",
     wallets: [...initiaWallets, ...keplrWallets],
     features: {
       faucet: {
@@ -47,7 +47,7 @@ export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
       gasPrice: {
         tokenPerGas: 0.151,
         denom:
-          "l2/aee375e9d0b181f0d9d3a49f9a3d1d6b05d62b0ac81f8c92b9282afa4213d884",
+          "l2/771d639f30fbe45e3fbca954ffbe2fcc26f915f5513c67a4a2d0bc1d635bdefd",
       },
       gasAdjustment: 1.5,
       maxGasLimit: 25_000_000,
@@ -56,14 +56,14 @@ export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
       isValidatorExternalLink: null,
     },
   },
-  "tomcat-1-sequencer": {
+  "minimove-1-sequencer": {
     tier: "sequencer",
     chain: "initia",
-    registryChainName: "blackwingtestnet",
-    prettyName: "Blackwing Testnet Sequencer",
-    lcd: "https://maze-rest-18bdff44-3aa4-425e-9bc0-06a2afa40af8.ase1-prod.newmetric.xyz",
-    rpc: "https://maze-rpc-18bdff44-3aa4-425e-9bc0-06a2afa40af8.ase1-prod.newmetric.xyz",
-    indexer: "https://tomcat-1-graphql.alleslabs.dev/v1/graphql",
+    registryChainName: "minimovetestnet",
+    prettyName: "Minimove",
+    lcd: "https://lcd.minimove-1.initia.xyz",
+    rpc: "https://rpc.minimove-1.initia.xyz",
+    indexer: "https://minimove-1-graphql.alleslabs.dev/v1/graphql",
     wallets: [...initiaWallets, ...keplrWallets],
     features: {
       faucet: {
@@ -95,7 +95,7 @@ export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
       gasPrice: {
         tokenPerGas: 0.151,
         denom:
-          "l2/aee375e9d0b181f0d9d3a49f9a3d1d6b05d62b0ac81f8c92b9282afa4213d884",
+          "l2/771d639f30fbe45e3fbca954ffbe2fcc26f915f5513c67a4a2d0bc1d635bdefd",
       },
       gasAdjustment: 1.5,
       maxGasLimit: 25_000_000,
@@ -479,6 +479,53 @@ export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
         tokenPerGas: 0.151,
         denom:
           "l2/afaa3f4e1717c75712f8e8073e41f051a4e516cd25daa82d948c4729388edefd",
+      },
+      gasAdjustment: 1.5,
+      maxGasLimit: 25_000_000,
+    },
+    extra: {
+      isValidatorExternalLink: null,
+    },
+  },
+  "glados-1": {
+    tier: "sequencer",
+    chain: "initia",
+    registryChainName: "milkywaytestnet",
+    prettyName: "Milkyway",
+    lcd: "https://lcd.testnet.milkyway.zone",
+    rpc: "https://rpc.testnet.milkyway.zone",
+    indexer: "",
+    wallets: [...initiaWallets, ...keplrWallets],
+    features: {
+      faucet: {
+        enabled: false,
+      },
+      wasm: {
+        enabled: true,
+        storeCodeMaxFileSize: 800_000,
+        clearAdminGas: 50_000,
+      },
+      move: {
+        enabled: false,
+      },
+      pool: {
+        enabled: false,
+      },
+      publicProject: {
+        enabled: true,
+      },
+      gov: {
+        enabled: false,
+      },
+      nft: {
+        enabled: false,
+      },
+    },
+    gas: {
+      gasPrice: {
+        tokenPerGas: 0.151,
+        denom:
+          "ibc/37A3FB4FED4CA04ED6D9E5DA36C6D27248645F0E22F585576A1488B8A89C5A50",
       },
       gasAdjustment: 1.5,
       maxGasLimit: 25_000_000,