+++ src/lib/pages/home/index.tsx
@@ -6,6 +6,7 @@ import { TierSwitcher } from "lib/components/TierSwitcher";
 
 import { HomeFull } from "./full";
 import { HomeLite } from "./lite";
+import { HomeSequencer } from "./sequencer";
 
 const Home = () => {
   const router = useRouter();
@@ -14,7 +15,13 @@ const Home = () => {
     if (router.isReady) track(AmpEvent.TO_OVERVIEW);
   }, [router.isReady]);
 
-  return <TierSwitcher full={<HomeFull />} lite={<HomeLite />} />;
+  return (
+    <TierSwitcher
+      full={<HomeFull />}
+      sequencer={<HomeSequencer />}
+      lite={<HomeLite />}
+    />
+  );
 };
 
 export default Home;