+++ src/lib/services/tx/sequencer.ts
@@ -1,8 +1,48 @@
 import axios from "axios";
 
-import { zBlockTxsResponseSequencer } from "../types";
-import type { Option } from "lib/types";
+import {
+  zBlockTxsResponseSequencer,
+  zTxsByHashResponseSequencer,
+  zTxsResponseSequencer,
+} from "../types";
+import type { BechAddr20, Option } from "lib/types";
 import { parseWithError } from "lib/utils";
+
+export const getTxsSequencer = async (
+  endpoint: string,
+  paginationKey: Option<string>,
+  limit: number
+) =>
+  axios
+    .get(`${endpoint}/indexer/tx/v1/txs`, {
+      params: {
+        "pagination.limit": limit,
+        "pagination.reverse": true,
+        "pagination.key": paginationKey,
+      },
+    })
+    .then(({ data }) => parseWithError(zTxsResponseSequencer, data));
+
+export const getTxsByAccountAddressSequencer = async (
+  endpoint: string,
+  address: BechAddr20,
+  paginationKey: Option<string>,
+  limit: number
+) =>
+  axios
+    .get(`${endpoint}/indexer/tx/v1/txs/by_account/${encodeURI(address)}`, {
+      params: {
+        "pagination.limit": limit,
+        "pagination.reverse": true,
+        "pagination.key": paginationKey,
+      },
+    })
+    .then(({ data }) => parseWithError(zTxsResponseSequencer, data));
+
+export const getTxsByHashSequencer = async (endpoint: string, txHash: string) =>
+  axios
+    .get(`${endpoint}/indexer/tx/v1/txs/${encodeURI(txHash)}`)
+    .then(({ data }) => parseWithError(zTxsByHashResponseSequencer, data));
 
 export const getTxsByBlockHeightSequencer = async (
   endpoint: string,