+++ src/lib/services/types/tx.ts
@@ -22,6 +22,7 @@ import {
   extractTxLogs,
   getActionMsgType,
   getMsgFurtherAction,
+  getTxBadges,
   parseTxHash,
   snakeToCamel,
 } from "lib/utils";
@@ -105,7 +106,7 @@ export type Tx = z.infer<typeof zTx>;
 
 const zEventAttribute = z.object({
   key: z.string(),
-  value: z.string(),
+  value: z.union([z.string(), z.null().transform(() => "")]),
   index: z.boolean().optional(),
 });
 
@@ -161,6 +162,17 @@ export const zTxsResponseItemFromLcd =
       type: msg["@type"],
     }));
 
+    const { isIbc, isOpinit } = messages.reduce(
+      (acc, msg, idx) => {
+        const current = getTxBadges(msg.type, logs[idx]);
+        return {
+          isIbc: acc.isIbc || current.isIbc,
+          isOpinit: acc.isOpinit || current.isOpinit,
+        };
+      },
+      { isIbc: false, isOpinit: false }
+    );
+
     return {
       hash: val.txhash,
       messages,
@@ -169,11 +181,12 @@ export const zTxsResponseItemFromLcd =
       height: Number(val.height),
       created: val.timestamp,
       success: val.code === 0,
+      isIbc,
+      isOpinit,
+      events: val.events,
       // TODO: implement below later
       actionMsgType: ActionMsgType.OTHER_ACTION_MSG,
       furtherAction: MsgFurtherAction.NONE,
-      isIbc: false,
-      isOpinit: false,
       isInstantiate: false,
     };
   });
@@ -189,6 +202,16 @@ export const zTxsByAddressResponseLcd = z
   }));
 export type TxsByAddressResponseLcd = z.infer<typeof zTxsByAddressResponseLcd>;
 
+export const zTxsResponseSequencer = z
+  .object({
+    txs: z.array(zTxsResponseItemFromLcd),
+    pagination: zPagination,
+  })
+  .transform((val) => ({
+    items: val.txs,
+    pagination: val.pagination,
+  }));
+
 export const zTxsByHashResponseLcd = z
   .object({
     tx_response: zTxsResponseItemFromLcd,
@@ -197,6 +220,21 @@ export const zTxsByHashResponseLcd = z
     items: [val.tx_response],
     total: 1,
   }));
+
+export const zTxsByHashResponseSequencer = z
+  .object({
+    tx: zTxsResponseItemFromLcd,
+  })
+  .transform((val) => ({
+    items: [val.tx],
+    pagination: {
+      total: val.tx ? 1 : 0,
+      nextKey: null,
+    },
+  }));
+export type TxsByHashResponseSequencer = z.infer<
+  typeof zTxsByHashResponseSequencer
+>;
 
 export const zTxByHashResponseLcd = z
   .object({